docs: move basic-starter logic into src directory
