+++ examples/alb/src/app.js
@@ -3,7 +3,6 @@ const express = require('express')
 const bodyParser = require('body-parser')
 const cors = require('cors')
 // const compression = require('compression')
-// const { getCurrentLambdaInvoke } = require(process.env.NODE_ENV === 'test' ? '../..' : 'aws-serverless-express')
 const { getCurrentLambdaInvoke } = require('aws-serverless-express')
 const app = express()
 const router = express.Router()
