+++ examples/basic-starter/src/lambda.js
