+++ examples/basic-starter/src/sam-logo.png
