+++ examples/basic-starter/src/views/index.pug
